Added ready notification to network.sh.
Also removed all endian flipping, using standard functions.

+++ descriptor_header.h
@@ -28,7 +28,7 @@ public:
 	DescriptorHeader(header_type type);
 	DescriptorHeader(MessageId &messageID, header_type type,
 			unsigned short time_to_live, unsigned short hops,
-			unsigned long payload_len);
+			uint32_t payload_len);
 	DescriptorHeader(in_port_t port, in_addr_t addr);
 	~DescriptorHeader();
 	const char *get_header();
@@ -36,7 +36,7 @@ public:
 	header_type get_header_type();
 	unsigned short get_time_to_live();
 	unsigned short get_hops();
-	unsigned long get_payload_len();
+	uint32_t get_payload_len();
 	in_port_t get_port();
 	in_addr_t get_addr();
 };
