+++ payload.h
@@ -21,28 +21,28 @@ using namespace std;
 class Payload {
 protected:
 	char *m_payload;
-	unsigned long m_payload_len;
+	uint32_t m_payload_len;
 public:
 	Payload();
 	virtual ~Payload();
 	const char *get_payload();
-	unsigned long get_payload_len();
+	uint32_t get_payload_len();
 };
 
 class Pong_Payload : public Payload {
 private:
 	in_port_t m_port;
 	in_addr_t m_ip_addr;
-	unsigned long m_files_shared;
-	unsigned long m_kilo_shared;
+	uint32_t m_files_shared;
+	uint32_t m_kilo_shared;
 public:
-	Pong_Payload(in_port_t port, in_addr_t ip_addr, unsigned long files_shared,
-			unsigned long kilo_shared);
+	Pong_Payload(in_port_t port, in_addr_t ip_addr, uint32_t files_shared,
+			uint32_t kilo_shared);
 	Pong_Payload(const char *payload);
 	in_port_t get_port();
 	in_addr_t get_ip_addr();
-	unsigned long get_files_shared();
-	unsigned long get_kilo_shared();
+	uint32_t get_files_shared();
+	uint32_t get_kilo_shared();
 };
 
 class Query_Payload : public Payload {
@@ -51,23 +51,23 @@ private:
 	string m_search;
 public:
 	Query_Payload(unsigned short speed, string search);
-	Query_Payload(const char *payload, unsigned long payload_len);
+	Query_Payload(const char *payload, uint32_t payload_len);
 	unsigned short get_speed();
 	string get_search();
 };
 
 class Result : public Payload {
 private:
-	unsigned long m_file_index;
-	unsigned long m_file_size;
+	uint32_t m_file_index;
+	uint32_t m_file_size;
 	string m_file_name;
 public:
-	Result(unsigned long file_index, unsigned long file_size,
+	Result(uint32_t file_index, uint32_t file_size,
 			string m_file_name);
-	Result(const char *result, unsigned long length);
+	Result(const char *result, uint32_t length);
 
-	unsigned long const get_file_index();
-	unsigned long const get_file_size();
+	uint32_t const get_file_index();
+	uint32_t const get_file_size();
 	string const get_file_name();
 };
 
@@ -76,13 +76,13 @@ private:
 	unsigned short m_num_hits;
 	in_port_t m_port;
 	in_addr_t m_ip_addr;
-	unsigned long m_speed;
+	uint32_t m_speed;
 	vector<Result> m_result_set;
 	char m_servent_id[16];
 public:
-	QueryHit_Payload(in_port_t port, in_addr_t ip_addr, unsigned long speed,
+	QueryHit_Payload(in_port_t port, in_addr_t ip_addr, uint32_t speed,
 			vector<Result> result_set, const char *servent_id);
-	QueryHit_Payload(const char *payload, unsigned long payload_len);
+	QueryHit_Payload(const char *payload, uint32_t payload_len);
 	unsigned short get_num_hits();
 	in_port_t get_port();
 	in_addr_t get_ip_addr();
@@ -93,31 +93,31 @@ class HTTPget_Payload : public Payload {
 private: 
 	string m_request;
 public:
-	HTTPget_Payload(unsigned long file_index, unsigned long file_size, string file_name);
-	HTTPget_Payload(const char *payload, unsigned long payload_len);
+	HTTPget_Payload(uint32_t file_index, uint32_t file_size, string file_name);
+	HTTPget_Payload(const char *payload, uint32_t payload_len);
 	string get_request() { return m_request; }
 };
 class HTTPok_Payload : public Payload {
 private:
 	string m_response;
 public:
-	HTTPok_Payload(unsigned long file_size);
-	HTTPok_Payload(const char *payload, unsigned long payload_len);
+	HTTPok_Payload(uint32_t file_size);
+	HTTPok_Payload(const char *payload, uint32_t payload_len);
 	string get_response() { return m_response; }
 };	
 
 class Push_Payload : public Payload {
 private:
 	char m_servent_id[16];
-	unsigned long m_file_index;
+	uint32_t m_file_index;
 	in_port_t m_port;
 	in_addr_t m_ip_addr;
 public:
-	Push_Payload(const char *servent_id, unsigned long file_index,
+	Push_Payload(const char *servent_id, uint32_t file_index,
 			in_port_t port, in_addr_t ip_addr);
 	Push_Payload(const char *payload);
 	const char *get_servent_id();
-	unsigned long get_file_index();
+	uint32_t get_file_index();
 	in_port_t get_port();
 	in_addr_t get_ip_addr();
 };
